evaluate if resource is namespaced before install
This commit has below fixes:
- Introduces new predicate w.r.t unstructured instances
- This predicate evaluates if the unstructured instance is
namespaced scope
- sparse pool related storageclass resource is added

Other changes:
- The unstructured.go file is split into unstructured.go &
resource.go
- installer.go file has been made more informative via comments
